Week4-Project-advance/requirements.txt
python>=3.6
tkinter
requests
json
Pillow
